+++ routes.py
@@ -6,6 +6,7 @@ routes = {
     "latest": "/posts/latest",
     "new_post": "/posts/new",
     "post_edit": "/posts/{}/edit",
+    "post_delete": "/posts/{}/delete",
     "login": "/users/login",
     "signup": "/users/signup",
     "switch": "/users/switch",
@@ -33,7 +34,6 @@ routes = {
     "admin_client_list": "/admin/clients/",
     "admin_client_sessions": "/admin/clients/{}/sessions",
     "admin_client_punish": "/admin/clients/{}/punish",
-    "admin_post_delete": "/admin/posts/{}/delete",
     "admin_posts_pending": "/admin/posts/pending",
     "admin_post_approval": "/admin/posts/{}/approval",
     "admin_comment_delete": "/admin/comments/{}/delete",
