+++ templates/posts/post.html
@@ -23,7 +23,8 @@
         {% if user %}
             <button id="like" data-id="{{ post.post_id }}" data-like="{{ not user.is_liked(post.post_id) }}"></button><br/>
             {% if user.user_id == post.author.user_id %}
-                <a href="{{ request.url + "\edit" }}">Edit Post...</a>
+                <a href="{{ routes["post_edit"].format(post.post_id) }}">Edit...</a>
+                <a href="{{ routes["post_delete"].format(post.post_id) }}">Delete...</a>
             {% endif %}
         {% endif %}
         <div id="post-content">{{ post.content }}</div>
@@ -33,11 +34,11 @@
             {% endfor %}
         </div>
         {% include 'comments.html' %}
-        {% if (user) and (user.role.value != "member") %}
+        {% if (user) and (user.role.value != "member") and (user.user_id != post.author.user_id) %}
             <div class="mod-menu">
                 <h2>{{ user.role.value.capitalize() }} Options</h2>
                 <ul>
-                    <li><a href="{{ routes["admin_post_delete"].format(post.post_id) }}">Delete</a></li>
+                    <li><a href="{{ routes["post_delete"].format(post.post_id) }}">Delete</a></li>
                 </ul>
             </div>
         {% endif %}
